Remember credentials
The app asks for the credentials after each cold start. The user likely wants to use the same login all the time, so it makes sense to prefill them. Feel free to not support remembering the password, but I personally don't care if an app does (especially in a case where you can't do much harm).